+++ Assets/Scripts/GamePlay/HumbleObjects/InjectorHumbleObject.cs
@@ -1,5 +1,4 @@
 using Source.GamePlay.Services;
-using Source.GamePlay.Services.Units;
 using Source.Shared.Controllers;
 using Source.Shared.Services;
 using Source.Shared.Utilities;
