+++ Assets/Scripts/GamePlay/HumbleObjects/UnitHumbleObject.cs
@@ -1,5 +1,5 @@
 using Source.GamePlay.HumbleObjects.Interfaces;
-using Source.GamePlay.Services.Units;
+using Source.GamePlay.Services;
 using System;
 using UnityEngine;
 using UnityEngine.AI;
