+++ Assets/Scripts/GamePlay/Services/GamePlayService.cs
@@ -1,5 +1,4 @@
 using Source.GamePlay.Services.Interfaces;
-using Source.GamePlay.Services.Units;
 using System.Collections.Generic;
 using System.Linq;
 using UnityEngine;
